Rename SocketType -> EngineType.
